[#32]refactor: 컴포넌트 폴더 내부 폴더화
페이지, 공용으로 폴더화
